Upgrade ExtUtils::Install to 1.52
git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@4369 8151f2b9-fde8-0310-94fd-f048d12aab9e

+++ inc/ExtUtils/Install.pm
@@ -38,11 +38,11 @@ ExtUtils::Install - install files from here to there
     
 =head1 VERSION
 
-1.50
+1.52
 
 =cut
 
-$VERSION = '1.50';
+$VERSION = '1.52';
 $VERSION = eval $VERSION;
 
 =pod
@@ -248,8 +248,9 @@ sub _unlink_or_rename { #XXX OS-SPECIFIC
     my ( $file, $tryhard, $installing )= @_;
 
     _chmod( 0666, $file );
-    unlink $file
-        and return $file;
+    my $unlink_count = 0;
+    while (unlink $file) { $unlink_count++; }
+    return $file if $unlink_count > 0;
     my $error="$!";
 
     _choke("Cannot unlink '$file': $!")
@@ -358,7 +359,8 @@ Abstract a -w check that tries to use POSIX::access() if possible.
     sub _have_write_access {
         my $dir=shift;
         unless (defined $has_posix) {
-            $has_posix= (!$Is_cygwin && eval 'local $^W; require POSIX; 1') || 0; 
+            $has_posix= (!$Is_cygwin && !$Is_Win32
+			 && eval 'local $^W; require POSIX; 1') || 0;
         }
         if ($has_posix) {
             return POSIX::access($dir, POSIX::W_OK());
@@ -403,9 +405,14 @@ sub _can_write_dir {
     my $path='';
     my @make;
     while (@dirs) {
-        $dir = File::Spec->catdir(@dirs);
-        $dir = File::Spec->catpath($vol,$dir,'') 
-                if defined $vol and length $vol;
+        if ($Is_VMS) {
+            $dir = File::Spec->catdir($vol,@dirs);
+        }
+        else {
+            $dir = File::Spec->catdir(@dirs);
+            $dir = File::Spec->catpath($vol,$dir,'')
+                    if defined $vol and length $vol;
+        }
         next if ( $dir eq $path );
         if ( ! -e $dir ) {
             unshift @make,$dir;
@@ -809,7 +816,7 @@ sub install { #XXX OS-SPECIFIC
     if ($pack{'write'}) {
         $dir = install_rooted_dir(dirname($pack{'write'}));
         _mkpath( $dir, 0, 0755, $verbose, $dry_run );
-        print "Writing $pack{'write'}\n";
+        print "Writing $pack{'write'}\n" if $verbose;
         $packlist->write(install_rooted_file($pack{'write'})) unless $dry_run;
     }
 
@@ -950,6 +957,13 @@ sub install_default {
   my $INST_SCRIPT = File::Spec->catdir($Curdir,'blib','script');
   my $INST_MAN1DIR = File::Spec->catdir($Curdir,'blib','man1');
   my $INST_MAN3DIR = File::Spec->catdir($Curdir,'blib','man3');
+
+  my @INST_HTML;
+  if($Config{installhtmldir}) {
+      my $INST_HTMLDIR = File::Spec->catdir($Curdir,'blib','html');
+      @INST_HTML = ($INST_HTMLDIR => $Config{installhtmldir});
+  }
+
   install({
            read => "$Config{sitearchexp}/auto/$FULLEXT/.packlist",
            write => "$Config{installsitearch}/auto/$FULLEXT/.packlist",
@@ -961,6 +975,7 @@ sub install_default {
            $INST_SCRIPT => $Config{installscript},
            $INST_MAN1DIR => $Config{installman1dir},
            $INST_MAN3DIR => $Config{installman3dir},
+	   @INST_HTML,
           },1,0,0);
 }
 
@@ -1039,7 +1054,7 @@ sub inc_uninstall {
     #warn join "\n","---",@dirs,"---";
     my $seen_ours;
     foreach $dir ( @dirs ) {
-        my $canonpath = File::Spec->canonpath($dir);
+        my $canonpath = $Is_VMS ? $dir : File::Spec->canonpath($dir);
         next if $canonpath eq $Curdir;
         next if $seen_dir{$canonpath}++;
         my $targetfile = File::Spec->catfile($canonpath,$libdir,$file);
@@ -1077,7 +1092,7 @@ sub inc_uninstall {
             eval {
                 die "Fake die for testing" 
                     if $ExtUtils::Install::Testing and
-                       File::Spec->canonpath($ExtUtils::Install::Testing) eq $targetfile;
+                       ucase(File::Spec->canonpath($ExtUtils::Install::Testing)) eq ucase($targetfile);
                 forceunlink($targetfile,'tryhard');
                 $results->{uninstall}{$targetfile} = $filepath;
                 1;
@@ -1224,7 +1239,8 @@ sub DESTROY {
         }
         $plural = $i>1 ? "all those files" : "this file";
         my $inst = (_invokant() eq 'ExtUtils::MakeMaker')
-                 ? ( $Config::Config{make} || 'make' ).' install UNINST=1'
+                 ? ( $Config::Config{make} || 'make' ).' install'
+                     . ( $Is_VMS ? '/MACRO="UNINST"=1' : ' UNINST=1' )
                  : './Build install uninst=1';
         print "## Running '$inst' will unlink $plural for you.\n";
     }
